+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Breaking changes
   `TransactionId` and use in `TxCodec.identifier`, `PostTxResponse` and
   `BcpTxQuery`.
 * @iov/bcp-types: `BcpTxQuery.tags` is now optional
+* @iov/bcp-types: Remove type `RecipientId`; use `Address` instead.
 * @iov/bns: `BnsConnection.postTx` now resolves before a transaction is in a
   block. The field `blockInfo` of its response can be used to track the
   transaction state.